+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - Address authentication for singlesig software wallets
+- Add lightning closing channels and sweep funds
 
 ## [4.0.10] - 2023-07-25